Integration: fix paddle graphics/collision mismatch
Prior to this commit, there existed a very egregious error that caused
the paddle locations to be swapped in collision detection vs. graphics.
This caused very confusing problems.  The issue is now fixed.

In the process, I added a simple LED output to signal when a paddle
collision has occurred.